% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{get_predictit}
\alias{get_predictit}
\title{Pull data from Predict It's API}
\usage{
get_predictit(category, subcategory = "",
  agent = "predictitR user, no message set")
}
\arguments{
\item{category}{overall categories eg group or ticker}

\item{subcategory}{specific markets you want to pull, eg 6 for US elections or  USPREZ for 2016 presidential market}

\item{agent}{a note for Predict It staff, indicating who you are and why you're pulling the data}
}
\value{
a dataframe with market data for each market
}
\description{
Pull data from Predict It's API
}
\examples{
For US election markets:
get_predictit(category="category", subcategory="6")

For 2016 election markets:
get_predictit("ticker", "USPREZ16")

For all markets:
get_predictit("all")

pull from US Electoral College markets, setting agent properly:
get_predictit(category="group", subcategory="75", agent="Jane Doe for research project")
}
