% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_functions.R
\name{list_of_markets}
\alias{list_of_markets}
\title{Get list of all current markets and their ticker symbol}
\usage{
list_of_markets(agent = "predictitr user, no message set")
}
\arguments{
\item{agent}{a note for Predict It staff, indicating who you are and why you're pulling the data}
}
\value{
a data frame containing every current market and their ticker symbol
}
\description{
Get list of all current markets and their ticker symbol
}
\examples{
list_of_markets()
list_of_market(agent="Jane Doe, pulling data for my senior thesis @ great university")
}
